Move `_determineRunningLocation` up to `AbstractExtensionService`
